closes #13 headers proxied to sub calls in batch handle batch processing moved to internal dispatching instead of outer requests
